@@ -22,45 +22,52 @@
     "email": "joachimwester@me.com",
     "url": "http://www.starcounter.com/"
   },
+  "main": "./dist/palindrom.node.js",
+  "browser": "./dist/palindrom.min.js",
+  "files": [
+    "dist/palindrom.node.js"
+  ],
   "licenses": [
     {
       "type": "MIT",
       "url": "http://www.opensource.org/licenses/MIT"
     }
   ],
-  "main": "./src/palindrom.js",
   "dependencies": {
-    "axios": "^0.15.3",
     "events": "^1.1.1",
     "fast-json-patch": "^2.0.7",
     "json-patch-ot": "^1.0.1",
     "json-patch-ot-agent": "2.0.0-rc.0",
     "jsonpatcherproxy": "^0.0.10",
+    "node-fetch": "^2.3.0",
     "url": "^0.11.0",
     "websocket": "^1.0.26"
   },
   "devDependencies": {
-    "babel-minify-webpack-plugin": "^0.2.0",
+    "@webcomponents/webcomponentsjs": "^1.3.3",
     "bluebird": "^3.5.0",
     "bluebird-retry": "^0.10.1",
-    "chai": "^4.0.0",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "colors": "^1.3.2",
-    "jasmine": "^2.99.0",
-    "json-loader": "^0.5.4",
-    "mocha": "^5.2.0",
+    "fetch-mock": "^7.3.1",
+    "mocha": "^6.1.4",
     "mock-socket": "6.0.4",
-    "moxios": "^0.3.0",
-    "polyserve": "^0.27.12",
     "saucelabs": "^1.5.0",
     "selenium-webdriver": "^3.3.0",
-    "sinon": "^2.1.0",
-    "webpack": "^3.12.0"
+    "serve": "^11.0.0",
+    "sinon": "^7.3.2",
+    "webpack": "^4.32.0",
+    "webpack-cli": "^3.2.3"
   },
   "scripts": {
     "version": "node ./bump-version.js && webpack && git add -A",
+    "serve": "serve",
+    "test": "webpack && mocha test/runner-node.js && node test/Sauce/Runner.js",
     "test-sauce": "webpack && node test/Sauce/Runner.js",
-    "test": "mocha test/runner.js",
-    "test-full": "mocha test/runner.js && webpack && node test/Sauce/Runner.js",
+    "test-node": "webpack && mocha test/runner-node.js",
+    "debug-tests-node": "mocha --inspect-brk test/runner-node.js",
+    "build-watch": "webpack  --watch",
     "build": "webpack"
   }
 }
